requirements.txt
google-adk==1.9.0
google-genai
a2a-sdk==0.3.0
flask
flask-cors
python-dotenv
requests
termcolor
nest-asyncio
uvicorn
aiohttp
